Reading matrix from file removes duplicates.
