+++ resolve/matrix/Utilities.cpp
@@ -46,88 +46,6 @@ namespace ReSolve
       real_type value_;
   };
 
-  /// @brief Helper class for COO matrix sorting
-  class CooTriplet
-  {
-    public:
-      CooTriplet() : rowidx_(0), colidx_(0), value_(0.0)
-      {}
-      ~CooTriplet()
-      {}
-      void setColIdx (index_type new_idx)
-      {
-        colidx_ = new_idx;
-      }
-      void setValue (real_type new_value)
-      {
-        value_ = new_value;
-      }
-      void set(index_type rowidx, index_type colidx, real_type value)
-      {
-        rowidx_ = rowidx;
-        colidx_ = colidx;
-        value_  = value;
-      }
-
-      index_type getRowIdx()
-      {
-        return rowidx_;
-      }
-      index_type getColIdx()
-      {
-        return colidx_;
-      }
-      real_type getValue()
-      {
-        return value_;
-      }
-
-      bool operator < (const CooTriplet& str) const
-      {
-        if (rowidx_ < str.rowidx_)
-          return true;
-
-        if ((rowidx_ == str.rowidx_) && (colidx_ < str.colidx_))
-          return true;
-
-        return false;
-      }
-
-      bool operator == (const CooTriplet& str) const
-      {
-        return (rowidx_ == str.rowidx_) && (colidx_ == str.colidx_);
-      }
-
-      CooTriplet& operator += (const CooTriplet t)
-      {
-        if ((rowidx_ != t.rowidx_) || (colidx_ != t.colidx_)) {
-          out::error() << "Adding values into non-matching triplet.\n";
-        }
-        value_ += t.value_;
-        return *this;
-      }
-
-      void print() const
-      {
-        // Add 1 to indices to restore indexing from MM format
-        std::cout << rowidx_ << " " << colidx_ << " " << value_ << "\n";
-      }
-
-    private:
-      index_type rowidx_{0};
-      index_type colidx_{0};
-      real_type value_{0.0};
-  };
-
-  void print_list(std::list<CooTriplet>& l)
-  {
-    // Print out the list
-    std::cout << "tmp list:\n";
-    for (CooTriplet& n : l)
-      n.print();
-    std::cout << "\n";
-  }
-
 
   namespace matrix
   {
@@ -229,9 +147,13 @@ namespace ReSolve
       // Convert to general CSR
       // Complexity O(NNZ)
       A_csr->setExpanded(true);
-      A_csr->setNnz(nnz_expanded_no_duplicates);
+      A_csr->setSymmetric(is_symmetric);
+      A_csr->setNnz(nnz); //nnz_expanded_no_duplicates);
       A_csr->setNnzExpanded(nnz_expanded_no_duplicates);
       A_csr->allocateMatrixData(memory::HOST);
+      if (memspace != memory::HOST) {
+        A_csr->allocateMatrixData(memspace);
+      }
       index_type* rows_csr = A_csr->getRowData(memory::HOST);
       index_type* cols_csr = A_csr->getColData(memory::HOST);
       real_type*  vals_csr = A_csr->getValues(memory::HOST);
@@ -266,9 +188,10 @@ namespace ReSolve
       rows_csr[csr_row_counter] = nnz_expanded_no_duplicates;
 
       // Move data to `memspace`
-      if (memspace != memory::HOST) {
-        A_csr->copyData(memspace);
-      }
+      // if (memspace != memory::HOST) {
+        // A_csr->copyData(memspace);
+        A_csr->updateData(rows_csr, cols_csr, vals_csr, memory::HOST, memspace);
+      // }
 
       // std::cout << "Size of tmp list = " << tmp.size() << "\n\n";
       // std::cout << "Rows counted = " << csr_row_counter << "\n";
