+++ tests/unit/matrix/MatrixIoTests.hpp
@@ -284,9 +284,9 @@ R"(%%MatrixMarket matrix coordinate real general
 % Generated by Re::Solve <https://github.com/ORNL/ReSolve>
 5 5 8
 1 1 1.000000000000000e+00
+1 4 6.000000000000000e+00
 2 2 1.050000000000000e+01
 3 3 1.500000000000000e-02
-1 4 6.000000000000000e+00
 4 2 2.505000000000000e+02
 4 4 -2.800000000000000e+02
 4 5 3.332000000000000e+01
@@ -310,12 +310,12 @@ R"(%%MatrixMarket matrix coordinate real general
 )";
 
   /// Matching COO matrix data as it is supposed to be read from the file
-  const std::vector<index_type> general_coo_matrix_rows_ = {0,1,2,0,3,3,3,4};
-  const std::vector<index_type> general_coo_matrix_cols_ = {0,1,2,3,1,3,4,4};
+  const std::vector<index_type> general_coo_matrix_rows_ = {0,0,1,2,3,3,3,4};
+  const std::vector<index_type> general_coo_matrix_cols_ = {0,3,1,2,1,3,4,4};
   const std::vector<real_type> general_coo_matrix_vals_ = { 1.000e+00,
+                                                            6.000e+00,
                                                             1.050e+01,
                                                             1.500e-02,
-                                                            6.000e+00,
                                                             2.505e+02,
                                                            -2.800e+02,
                                                             3.332e+01,
